S302T300-136 #time 50m ?implement ?test Get Products from Catalogue -More research on mocking and testing service/dao layer -Refactored testing to remove redundant code

+++ backend/src/main/java/com/navbara_pigeons/wasteless/dao/ProductDao.java
@@ -8,6 +8,6 @@ public interface ProductDao {
 
     void saveProduct(Product product);
 
-    List<Product> getProducts(int businessId);
+    List<Product> getProducts(long businessId);
 
 }
